Refactorizacion de codigo y limpieza

+++ my-back/src/main/java/utez/edu/mx/crochetifyBack/services/cart/CartServiceImp.java
@@ -148,7 +148,7 @@ public class CartServiceImp implements CartService {
                                                         - (cartProduct.getQuantity() * currentStock.getPrice());
                                         currentCart.setTotal(updatedTotal);
                                 } else {
-                                        int difference = request.getQuantity() - cartProduct.getQuantity();
+                                        Long difference = request.getQuantity() - cartProduct.getQuantity();
 
                                         cartProduct.setQuantity(request.getQuantity());
